requirements/base.txt
django==4.1.7
psycopg2-binary==2.9.5
python-environ==0.4.54
django-cors-headers==3.13.0
django-jazzmin==2.6.0
drf-yasg==1.21.5
djangorestframework==3.14.0
django-filter==22.1
Pillow==9.4.0
django-rosetta==0.9.8